refactor(Source): Several changes
Working in documentation (WIP)

Created CommandExecutor in
ApplicationState to reduce ApplicationState dependency on external
libraries and centralize maintenance in a proper file

File
reorganization
