+++ src/lib/components/application-state/enums/environment.enum.ts
@@ -1,29 +1,36 @@
 /**
- * Enum to represent the different environments where the application can run.
+ * @enum
+ * @group Enhancers
+ * @category ApplicationState
+ * @description Enumerates the possible environments where the application can run.
+ * @example
+ * ```typescript
+ * const currentEnvironment = Environment.DEVELOPMENT;
+ * ```
  */
 export enum Environment {
   /**
-   * Indicates the development environment, usually local and for debugging.
+   * @description Indicates the development environment, usually local and for debugging.
    */
   DEVELOPMENT = 'development',
 
   /**
-   * Indicates the testing environment, used for automated tests.
+   * @description Indicates the testing environment, used for automated tests.
    */
   TESTING = 'testing',
 
   /**
-   * Indicates the staging environment, a pre-production setup.
+   * @description Indicates the staging environment, a pre-production setup.
    */
   STAGING = 'staging',
 
   /**
-   * Indicates the production environment, where the live application runs.
+   * @description Indicates the production environment, where the live application runs.
    */
   PRODUCTION = 'production',
 
   /**
-   * In case the environment is unknown.
+   * @description In case the environment is unknown.
    */
   UNKNOWN = 'unknown',
 }
