+++ src/lib/components/application-state/index.ts
@@ -1,3 +1,11 @@
-export * from './enhancers/application-state.enhancer';
+/**
+ * @category ApplicationState
+ * @description Manages the state-related logic of the application. Includes enhancers, interfaces, and enumerations related to application state.
+ * @example
+ * ```typescript
+ * import * as ApplicationState from './ApplicationState';
+ * ```
+ */
+export * from './application-state.enhancer';
 export * from './interfaces/application-state.interface';
 export * from './enums/environment.enum';
