--- src/lib/components/error-analysis/enhancers/error-analysis.enhancer.ts
@@ -1,115 +1,0 @@
-import ErrorStackParser from 'error-stack-parser';
-
-import { StackFrame } from '../../../shared/types';
-
-import { ErrorAnalysisInterface } from '../interfaces/error-analysis.interface';
-
-/**
- * @class ErrorAnalysisEnhancer
- *
- * The ErrorAnalysisEnhancer class enhances an error object with details
- * like file information, line number, column number, etc.
- */
-export class ErrorAnalysisEnhancer implements ErrorAnalysisInterface {
-  /**
-   * @private
-   * @type {Error | undefined }
-   *
-   * Holds the original error object.
-   */
-  private _originalError: Error | undefined = undefined;
-
-  /**
-   * @private
-   * @type {Array<StackFrame>}
-   *
-   * Holds the parsed stack trace details.
-   */
-  private _parsedStack: Array<StackFrame> = [];
-
-  private static _stackCache: Map<Error, Array<StackFrame>> = new Map();
-
-  /**
-   * @constructor
-   *
-   * Initializes a new instance of the ErrorAnalysisEnhancer class.
-   */
-  constructor() {}
-
-  /**
-   * @public
-   * @method setOriginalError
-   * @param {Error} originalError - The original error object.
-   * @returns {this} - Returns the instance for chaining.
-   *
-   * Sets the original error and triggers extraction of error information.
-   */
-  public setOriginalError(originalError: Error): this {
-    this._originalError = originalError;
-    this._extractErrorInfo();
-    return this;
-  }
-
-  /**
-   * @public
-   * @method originalError
-   * @returns {Error | undefined } - Returns the original error object.
-   */
-  public get originalError(): Error | undefined {
-    return this._originalError;
-  }
-
-  /**
-   * @private
-   * @method _extractErrorInfo
-   * @returns {this} - Returns the instance for chaining.
-   *
-   * Extracts detailed error information from the stack trace.
-   */
-  private _extractErrorInfo(): this {
-    if (ErrorAnalysisEnhancer._stackCache.has(this._originalError!)) {
-      this._parsedStack = ErrorAnalysisEnhancer._stackCache.get(
-        this._originalError!,
-      )!;
-      return this;
-    }
-
-    const errStack = this._originalError?.stack ?? 'unknown';
-
-    if (errStack === 'unknown') {
-      return this;
-    }
-
-    const stackFrames = ErrorStackParser.parse(this._originalError!);
-
-    this._parsedStack = stackFrames.map(
-      ({ functionName, fileName, lineNumber, columnNumber }) => ({
-        functionName: functionName || 'unknown',
-        fileName: fileName || 'unknown',
-        lineNumber: lineNumber ?? -1,
-        columnNumber: columnNumber ?? -1,
-        typeName: functionName?.split('.')[0] ?? 'unknown',
-      }),
-    );
-    ErrorAnalysisEnhancer._stackCache.set(
-      this._originalError!,
-      this._parsedStack,
-    );
-
-    return this;
-  }
-
-  /**
-   * @public
-   * @method parsedStack
-   * @returns {Array<StackFrame>} - Returns the parsed stack trace details.
-   *
-   * Getter for the parsed stack trace details.
-   */
-  public get parsedStack(): Array<StackFrame> {
-    if (this._parsedStack === null) {
-      this._extractErrorInfo();
-    }
-    return this._parsedStack!;
-  }
-}
