+++ src/lib/components/error-analysis/index.ts
@@ -1,3 +1,3 @@
-export * from './enhancers/error-analysis.enhancer';
+export * from './error-analysis.enhancer';
 
 export * from './interfaces/error-analysis.interface';
