+++ src/lib/components/filter/index.ts
@@ -1,3 +1,3 @@
-export * from './utilities/filter.utility';
+export * from './filter.utility';
 
 export * from './interfaces/filter.interface';
