+++ src/lib/components/identifiers/enums/severity.enum.ts
@@ -1,24 +1,37 @@
 /**
- * Enum to represent the severity level of an error.
+ * @enum Severity
+ * @group Enhancers
+ * @category Identifiers
+ *
+ * @example
+ * ```typescript
+ * const sev = Severity.HIGH;
+ * ```
+ *
+ * [[include:enums/severity.md]]
  */
 export enum Severity {
   /**
-   * Indicates a low-impact error that generally doesn't affect the system's functionality.
+   * @description Indicates a low-impact error that generally doesn't affect the system's overall functionality.
+   * @usecase Minor UI glitches, log-only issues, non-critical background task failures.
    */
   LOW = 'low',
 
   /**
-   * Indicates a moderate-impact error that may affect some minor functionalities.
+   * @description Indicates a moderate-impact error that may affect some minor functionalities.
+   * @usecase Invalid user input, API rate-limiting, partial feature malfunction.
    */
   MEDIUM = 'medium',
 
   /**
-   * Indicates a high-impact error that affects critical functionalities but doesn't halt the entire system.
+   * @description Indicates a high-impact error that affects critical functionalities but doesn't halt the entire system.
+   * @usecase Database access issues, missing critical resources, broken authentication.
    */
   HIGH = 'high',
 
   /**
-   * Indicates a critical-impact error that demands immediate attention, potentially causing system outage.
+   * @description Indicates a critical-impact error that demands immediate attention, potentially causing system outage.
+   * @usecase Server crashes, data corruption, security breaches.
    */
   CRITICAL = 'critical',
 }
