+++ src/lib/components/serializers/index.ts
@@ -1,3 +1,3 @@
-export * from './utilities/serializers.utility';
+export * from './serializers.utility';
 
 export * from './interfaces/serializers.interface';
