+++ src/lib/components/serializers/interfaces/serializers.interface.ts
@@ -2,7 +2,10 @@ import { JsonReplacer } from '../../../shared/types';
 
 /**
  * @interface Serializers
+ * @group Utilities
+ * @category Serializers
  *
+ * @description
  * The Serializers interface defines the contract for classes that need to serialize
  * out in JSON, CSV, XML and YAML.
  *
