+++ src/lib/components/system-context/index.ts
@@ -1,3 +1,3 @@
-export * from './enhancers/system-context.enhancer';
+export * from './system-context.enhancer';
 
 export * from './interfaces/system-context.interface';
