+++ src/lib/components/user-info/index.ts
@@ -1,3 +1,3 @@
-export * from './enhancers/user-info.enhancer';
+export * from './user-info.enhancer';
 
 export * from './interfaces/user-info.interface';
