+++ src/lib/index.ts
@@ -1,3 +1,5 @@
 export * from './core';
 
 export * from './components';
+
+export * from './shared';
