+++ src/lib/shared/types/json-replacer.type.ts
@@ -1,11 +1,14 @@
 /**
  * @typedef JsonReplacer
+ * @group Utilities
+ * @category Serializers
  *
+ * @description
  * A type representing a function that takes a key and value as arguments,
  * and returns a value. Typically used for customizing JSON stringification.
  *
- * @param {string} key - The key in the JSON object.
- * @param {any} value - The corresponding value.
+ * @param key - The key in the JSON object.
+ * @param value - The corresponding value.
  *
  * @returns {any} The new or modified value.
  */
