+++ src/lib/shared/types/query-params.type.ts
@@ -1,6 +1,9 @@
 /**
  * @type {object} QueryParams
+ * @group Enhancers
+ * @category HttpStatus
  *
+ * @description
  * Represents a single key-value pair for a query parameter in a URL.
  *
  * @property {string} key - The query parameter name.
