+++ src/lib/shared/validators/category.validator.ts
@@ -2,6 +2,8 @@ import z from 'zod';
 import { Category } from '../../components/identifiers';
 
 /**
- * Validator for Category Enum.
+ * @group Enhancers
+ * @category Identifiers
+ * @description Validator for Category Enum.
  */
 export const ValidCategory = z.nativeEnum(Category);
