+++ src/lib/shared/validators/http-methods.validator.ts
@@ -3,6 +3,8 @@ import z from 'zod';
 import { HttpMethods } from '../../components/http-status';
 
 /**
- * Validator for HttpMethods Enum.
+ * @group Ehancers
+ * @category HttpStatus
+ * @description Validator for HttpMethods Enum.
  */
 export const ValidHttpMethods = z.nativeEnum(HttpMethods);
