+++ src/lib/shared/validators/http-status-codes.validator.ts
@@ -3,6 +3,8 @@ import z from 'zod';
 import { HttpStatusCodes } from '../../components/http-status';
 
 /**
- * Validator for HttpStatusCodes Enum.
+ * @group Enhancers
+ * @category HttpStatus
+ * @description Validator for HttpStatusCodes Enum.
  */
 export const ValidHttpStatusCodes = z.nativeEnum(HttpStatusCodes);
