+++ src/lib/shared/validators/ip.validator.ts
@@ -1,6 +1,8 @@
 import z from 'zod';
 
 /**
- * Validates that a string is a proper IP address.
+ * @group Enhancers
+ * @category HttpStatus
+ * @description Validates that a string is a proper IP address.
  */
 export const ValidIP = z.string().ip();
