+++ src/lib/shared/validators/number.validator.ts
@@ -1,6 +1,9 @@
 import z from 'zod';
 
 /**
- * Validates that a number is an integer and greater than 0.
+ * @group Enhancers
+ * @category HttpStatus
+ * @category Identifiers
+ * @description Validates that a number is an integer and greater than 0.
  */
 export const ValidNumber = z.number().int().gt(0);
