+++ src/lib/shared/validators/severity.validator.ts
@@ -2,6 +2,8 @@ import z from 'zod';
 import { Severity } from '../../components/identifiers';
 
 /**
- * Validator for Severity Enum.
+ * @group Enhancers
+ * @category Identifiers
+ * @description Validator for Severity Enum.
  */
 export const ValidSeverity = z.nativeEnum(Severity);
