+++ src/lib/shared/validators/url.validator.ts
@@ -1,6 +1,8 @@
 import z from 'zod';
 
 /**
- * Validates that a string is a proper URL.
+ * @group Enhancers
+ * @category HttpStatus
+ * @description Validates that a string is a proper URL.
  */
 export const ValidURL = z.string().url();
